% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{application_spec}
\alias{application_spec}
\title{Get Application}
\usage{
application_spec(filename)
}
\arguments{
\item{filename}{A filename.}
}
\value{
The Application.
}
\description{
Identify the macro application.
}
\keyword{internal}
