% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{arg_spec}
\alias{arg_spec}
\title{Get VBS Casting Function}
\usage{
arg_spec(arg)
}
\arguments{
\item{arg}{An R primitive.}
}
\value{
The VBS casting function to use.
}
\description{
Identify the VBS casting function to use for an R object.
}
\keyword{internal}
