% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bat_echolocation_calls}
\alias{bat_echolocation_calls}
\title{Bat scholocation call recordings}
\format{
A data frame with 1816 observations of 4 variables:
\describe{
\item{bat}{The individual bat identifier.}
\item{species}{The bat species identifier.}
\item{family}{The species family identifier.}
\item{call_recording}{A list of calls recorded at a sampling rate of 500 kHz.}
\item{fundamental_frequency_contour}{A list of estimates for the fundamental
frequency taken from the bat echolocation call recording and the associated time
for the estimate within the call recording.}
\item{duration}{The time duration of the echolocation call. Reported in
milliseconds.}
\item{harmonic_order}{The number of harmonics in the echolocation call.}
\item{dominant_harmonic}{The harmonic carrying the most energy.}
}
}
\source{
{Bat echolocation call identification for biodiversity monitoring: a probabilistic approach - https://doi.org/10.1111/rssc.12217}

{Bayesian Ancestral Reconstruction for Bat Echolocation - https://ethos.bl.uk/OrderDetails.do?uin=uk.bl.ethos.828108}
}
\usage{
bat_echolocation_calls
}
\description{
A set of 1805 echolocation call recordings collected from 449 individual bats
across 22 species and 5 families. Calls were recorded at a sampling rate of
500 kHz. The original data set is described by Stathopolous et al. (2017)
(see link below). Additional features have been extracted from echolocation
calls allow comparative analysis. These feature are based on a harmonic model
for bat echolocation calls and the feature extraction procedure is described
in Meagher (2020) (see second link below)
}
\keyword{datasets}
