% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bat_phylogeny}
\alias{bat_phylogeny}
\title{Bat Phylogeny}
\format{
An object of class \code{phylo} of length 4.
}
\source{
{https://link.springer.com/article/10.1007/s10914-016-9363-8}
}
\usage{
bat_phylogeny
}
\description{
A phylogenetic tree for 22 species of Mexican bat transcribed from the
Maximum a posteriori phylogeny inferred by Amador et al. (2018) (see link
below).
}
\keyword{datasets}
