% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_plvm.R
\name{cavi_plvm}
\alias{cavi_plvm}
\title{CAVI for the PLVM}
\usage{
cavi_plvm(
  plvm_list,
  tol = 1e-06,
  max_iter = 1000,
  n_samples = 1000,
  random_seed = NULL,
  progress_bar = TRUE,
  perform_checks = TRUE
)
}
\arguments{
\item{plvm_list}{A list defining the PLVM generated by `initialise_plvm`.}

\item{tol}{A positive real-valued scalar. The required reducting in the ELBO to define convergence,}

\item{max_iter}{A positive integer. The maximum number of iterations that the optimisation procedure swill run.}

\item{n_samples}{A positive integer. The number of samples drawn to esimated expected auxiliary trait values.}

\item{random_seed}{The seed for random number generation.}

\item{progress_bar}{Logical. Should a progress bar be displayed?}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
The optimised PLVM and ELBO at each iteration.
}
\description{
Coordinate Ascent Variational Inference for the Phylogenetic Latent Variable Model
}
\seealso{
initialise_plvm
}
