% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_trait_metadata.R
\name{check_trait_type}
\alias{check_trait_type}
\title{Check trait type specification}
\usage{
check_trait_type(manifest_trait, name, type, levels)
}
\arguments{
\item{manifest_trait}{Variable(s) from a daframe. The manifest trait
recordings to be checked.}

\item{name}{A character scalar. The name of the trait to be checked.}

\item{type}{A caracter scalar. One of \eqn{c("ord", "nom", "con", "fvt")}.
The type of trait to be checked.}

\item{levels}{A scalar. The number of levels associated with a discrete manifest
trait. NA when traits are continuous.}
}
\value{
If the conditions are not met then returns an error message.
}
\description{
Check that specified trait type metadata matches the dataframe of manifest
traits. This provides a non-exhaustive series of checks.
}
