% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{collection_spec}
\alias{collection_spec}
\title{Get Collection Type}
\usage{
collection_spec(application)
}
\arguments{
\item{application}{An application, e.g. "Word" or "Excel".}
}
\value{
The collection type.
}
\description{
Identify the collection type.
}
\keyword{internal}
