% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_trait_metadata.R
\name{compare_metadata_to_df}
\alias{compare_metadata_to_df}
\title{Compare Metadata to Manifest Traits}
\usage{
compare_metadata_to_df(
  n_traits,
  manifest_trait_df,
  trait_names,
  trait_type,
  trait_levels,
  manifest_trait_index
)
}
\arguments{
\item{n_traits}{A natural number. The number of manifest traits.}

\item{manifest_trait_df}{A data frame. Observed manifest traits.}

\item{trait_names}{A P-dimensional character vector. The manifest trait
names.}

\item{trait_type}{A P-dimensional vector of unordered factors belonging to
one of 4 levels. The trait type corresponding to each of the named manifest
traits. Ordinal traits are denoted \eqn{ord}, nominal are \eqn{nom},
scalar-valued continuous are \eqn{con}, and function-valued traits are
\eqn{fvt}.}

\item{trait_levels}{A P-dimensional vector. The number of levels associated with each
manifest trait. NA for continuous traits.}

\item{manifest_trait_index}{A P-dimensional list. columns of
\eqn{manifest_trait_df} associated with each trait. List elements will be
scalar valued for ordinal, nominal, and scalar-valued continuous traits.
Function-valued traits are multi-variate and indexed by a vector of values.}
}
\value{
Reports error messages if metadata and dataframe do not match.
}
\description{
Compare the supplied meta data to that implicit in a data frame of manifest
traits.
}
