% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{compute_auxiliary_trait_elbo}
\alias{compute_auxiliary_trait_elbo}
\title{Auxiliary Trait ELBO}
\usage{
compute_auxiliary_trait_elbo(
  manifest_trait_df,
  metadata,
  auxiliary_traits,
  loading_expectation,
  latent_trait_expectation,
  precision,
  loading_outer_expectation,
  latent_trait_outer_expectation,
  n_samples = 1000,
  random_seed = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{manifest_trait_df}{The data frame of manifest traits. If included
allows for a non-exhaustive series of checks comparing metadata provided to
that implicit in the manifest traits.}

\item{metadata}{A data frame. Contains all the metadata required to map a set
of manifest traits to the PLVM auxiliary traits..}

\item{auxiliary_traits}{An NxD' matrix of real numbers. The auxiliary traits.}

\item{loading_expectation}{A D'xL matrix of real numbers, The expected
loading matrix.}

\item{latent_trait_expectation}{An NxL matrix of real values. The expected
individual specific latent traits.}

\item{precision}{A P-dimensional vector of positive real values. The
precision with which auxiliary traits are observed.}

\item{loading_outer_expectation}{A LxLxD' array. The expected outer products
of the expected loading matrix.}

\item{latent_trait_outer_expectation}{A LxLxN array. The expected outer
product of individual specific latent traits.}

\item{n_samples}{A positive integer. The number of independent samples drawn
to obtain the Monte Carlo estimate.}

\item{random_seed}{A single value, interpreted as an integer, or NULL.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A scalar. The contribution of Auxiliary traits to the ELBO.
}
\description{
Compute the contribution of auxiliary traits to the Evidence Lower Bound
(ELBO) of a PLVM given the approximate posterior distribution for loadings
and latent traits
}
