% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{compute_continuous_auxiliary_trait_elbo}
\alias{compute_continuous_auxiliary_trait_elbo}
\title{Continuous Trait ELBO}
\usage{
compute_continuous_auxiliary_trait_elbo(
  auxiliary_trait,
  loading_expectation,
  latent_trait_expectation,
  precision,
  loading_outer_expectation,
  latent_trait_outer_expectation,
  perform_checks = TRUE
)
}
\arguments{
\item{auxiliary_trait}{Either an N-dimensional vector or a NxD' matrix of
real values. The auxiliary trait measurements associated with a single
scalar- or function-valued trait.}

\item{loading_expectation}{Either a L-dimensional vector or D'xL matrix of
real numbers, The row(s) of the expected loading matrix corresponding to
the continuous trait.}

\item{latent_trait_expectation}{An NxL matrix of real values. The expected
individual specific latent traits.}

\item{precision}{A positive real valued scalar. The precision with which the
auxiliary trait is observed.}

\item{loading_outer_expectation}{A LxL matrix or LxLxD' array. The expected
outer product(s) for the row(s) of the expected loading matrix
corresponding to the continuous trait.}

\item{latent_trait_outer_expectation}{A LxLxN array. The expected outer
product of individual specific latent traits.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A scalar. The contribution to the ELBO made by the continuous trait.Ò
}
\description{
Compute the contribution of a continuous trait to the Evidence Lower Bound
(ELBO) of a PLVM given the approximate posterior distribution for loadings
and latent traits
}
