% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_latent_traits.R
\name{compute_individual_specific_latent_trait_expectation}
\alias{compute_individual_specific_latent_trait_expectation}
\title{Compute Individual-specific Latent Trait Expectation}
\usage{
compute_individual_specific_latent_trait_expectation(
  auxiliary_trait,
  loading,
  taxon_specific_latent_trait,
  precision_vector,
  individual_specific_latent_trait_precision,
  within_taxon_amplitude,
  perform_checks = TRUE
)
}
\arguments{
\item{auxiliary_trait}{A D-dimensional vector of real values. The
individual-specific auxiliary trait.}

\item{loading}{A DxL dimensional matrix of real values. The PLVM loading.}

\item{taxon_specific_latent_trait}{An L-dimensional vector of real values.
The taxon-specific latent trait values.}

\item{precision_vector}{A D-dimensional vector of positive
real values The precision parameters associated with each of the D
auxiliary traits.}

\item{individual_specific_latent_trait_precision}{A symmetric  LxL matrix of
real values. The precision associated with individual-specific latent
traits.}

\item{within_taxon_amplitude}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The within-taxon variation parameters for
each of the L latent traits.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An L-dimensional vector of real values. The expected value of the
  individual-specific latent trait under the approximate posterior
  distributiÒon
}
\description{
Computes the expected value of the individual-specific latent trait  within
the PLVM under the approximate posterior distribution.
}
