% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_latent_traits.R
\name{compute_individual_specific_latent_trait_precision}
\alias{compute_individual_specific_latent_trait_precision}
\title{Compute Individual-specific Latent Trait Precision}
\usage{
compute_individual_specific_latent_trait_precision(
  precision_vector,
  loading_outer_product_expectation,
  within_taxon_amplitude,
  perform_checks = TRUE
)
}
\arguments{
\item{precision_vector}{A D-dimensional vector of positive
real values The precision parameters associated with each of the D
auxiliary traits.}

\item{loading_outer_product_expectation}{An \eqn{LxLxD} dimensional array of
real values. The expected outer product of the loading under the
approximate posterior distribution.}

\item{within_taxon_amplitude}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The within-taxon variation parameters for
each of the L latent traits.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A symmetric LxL matrix of real values. The precision matrix of the
  individual-specific latent traits under the approximate posterior
  distribution.
}
\description{
Computes the precision matrix for the approximate posterior distribution of
individual-specific latent traits within the PLVM.
}
