% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_latent_traits.R
\name{compute_internal_taxon_specific_latent_trait_precision}
\alias{compute_internal_taxon_specific_latent_trait_precision}
\title{Compute Internal Taxon-specific Latent Trait Precision}
\usage{
compute_internal_taxon_specific_latent_trait_precision(
  child_taxa_conditional_expectation_weights,
  child_taxa_conditional_standard_deviations,
  conditional_standard_deviation,
  perform_checks = TRUE
)
}
\arguments{
\item{child_taxa_conditional_expectation_weights}{An L-dimensional vector of
real values on the interval \eqn{[0, 1]}. The weight of the taxon-specific latent trait
values in the conditional distribution of the latent
traits for its child taxa.}

\item{child_taxa_conditional_standard_deviations}{An L-dimensional vector of
real values on the interval \eqn{[0, 1]}. The standard deviation of the taxon-specific
latent traits of child taxa conditional on the parent taxon.}

\item{conditional_standard_deviation}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The standard deviation of the taxon-specific
latent trait conditional on its parent taxon.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An L-dimensional vector of positive real values.
}
\description{
Computes the precision of taxon-specific latent traits at internal nodes of
the phylogeny under the approximate posterior distribution for the PLVM.
}
