% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_loading.R
\name{compute_loading_elbo}
\alias{compute_loading_elbo}
\title{Loading ELBO}
\usage{
compute_loading_elbo(
  loading_expectation,
  loading_row_covariance,
  ard_precision,
  loading_prior_correlation_log_det = NULL,
  inv_loading_prior_correlation = NULL,
  loading_prior_correlation = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{loading_expectation}{A D'xL matrix of real numbers, The expected
loading matrix.}

\item{loading_row_covariance}{A LxLxD' array. The covariance of each row of
the loading matrix under the approximate posterior distribution.}

\item{ard_precision}{An L-dimensional vector of real numbers. The ARD
precision hyper-parameters on the columns of the loading matrix.}

\item{loading_prior_correlation_log_det}{A real valued scalar. The log
determinant of the loading prior correlation matrix.}

\item{inv_loading_prior_correlation}{A D'xD' matrix. The inverse of the
loading prior correlation matrix.}

\item{loading_prior_correlation}{A D'xD' matrix. The loading prior
correlation matrix.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A real valued scalar. The contribution of the loading matrix to the
  ELBO under the approximate posterior distribution.
}
\description{
Compute the contribution of the loading matrix Evidence Lower Bound (ELBO) of
a PLVM given the approximate posterior distribution for loadings, the prior
correlation matrix, and ARD precision hyperparameters.
}
