% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_loading.R
\name{compute_loading_row_precision}
\alias{compute_loading_row_precision}
\title{Compute precision for a row of the Loading Matrix}
\usage{
compute_loading_row_precision(
  total_individual_specific_latent_trait_outer_product_expectation,
  precision,
  ard_precision,
  scaled_conditional_row_variance,
  perform_checks = TRUE
)
}
\arguments{
\item{total_individual_specific_latent_trait_outer_product_expectation}{An
LxL matrix of real values. The sum of expected outer products for each of
the individual specific latent traits under the approximate posterior
within CAVI for the PLVM.}

\item{precision}{A vector length \eqn{n_traits} taking positive real values.
The Precision with which auxiliary traits are observed. Precision
parameters for discrete traits are fixed at 1.}

\item{ard_precision}{An L-dimensional vector of real numbers. The initial ARD
precision hyper-parameters on the loadings.}

\item{scaled_conditional_row_variance}{A positive real-valued scalar. The
variance of the loading in each row conditioned on the loading at every
other row before allowing for the ARD precision hyper-parameter.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An LxL matrix of real values. The precision associated with a row of
  the loading matrix under the approximate distribution within CAVI for a
  PLVM.
}
\description{
Computes the Gaussian precision associated with a row of the loading matrix
under the approximate posterior within Coordinate Ascent Variational
Inference (CAVI) for a Phylogenetic Latent Variable Model (PLVM).
}
