% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_loading.R
\name{compute_loading_row_precision_list}
\alias{compute_loading_row_precision_list}
\title{Compute precision for each row of the Loading Matrix}
\usage{
compute_loading_row_precision_list(
  total_individual_specific_latent_trait_outer_product_expectation,
  precision_vector,
  ard_precision,
  scaled_conditional_row_variance_vector,
  perform_checks = TRUE
)
}
\arguments{
\item{total_individual_specific_latent_trait_outer_product_expectation}{An
LxL matrix of real values. The sum of expected outer products for each of
the individual specific latent traits under the approximate posterior
within CAVI for the PLVM.}

\item{precision_vector}{A D'-dimensional vector of positive real values. The
precision along each dimension of the auxiliary traits.}

\item{ard_precision}{An L-dimensional vector of real numbers. The initial ARD
precision hyper-parameters on the loadings.}

\item{scaled_conditional_row_variance_vector}{A D'-dimensional vector of
positive real values. The variance of the loading in each row conditioned
on the loading at every other row before allowing for the ARD precision
hyper-parameter.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A list of length D' containing LxL matrices of real values. The
  precision associated with each row of the loading matrix under the
  approximate distribution within CAVI for a PLVM.
}
\description{
Computes the Gaussian precision associated with each row of the loading
matrix under the approximate posterior within Coordinate Ascent Variational
Inference (CAVI) for a Phylogenetic Latent Variable Model (PLVM). Stores the
resulting D' precision matrices as a list.
}
