% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{compute_nominal_auxiliary_trait_elbo}
\alias{compute_nominal_auxiliary_trait_elbo}
\title{Nominal Trait ELBO}
\usage{
compute_nominal_auxiliary_trait_elbo(
  y,
  n_samples = 1000,
  random_seed = NULL,
  loading_expectation,
  latent_trait_expectation,
  loading_outer_expectation,
  latent_trait_outer_expectation,
  perform_checks = TRUE
)
}
\arguments{
\item{y}{An N-dimensional vector of unordered factors with K levels. A nominal
manifest trait.}

\item{n_samples}{A positive integer. The number of independent samples drawn
to obtain the Monte Carlo estimate.}

\item{random_seed}{A single value, interpreted as an integer, or NULL.}

\item{loading_expectation}{A KxL matrix of real numbers, The row of
the expected loading matrix corresponding to the nominal trait.}

\item{latent_trait_expectation}{An NxL matrix of real values. The expected
individual specific latent traits.}

\item{loading_outer_expectation}{A LxLxK array. The expected outer products for
the row of the expected loading matrix corresponding to the nominal trait.}

\item{latent_trait_outer_expectation}{A LxLxN array. The expected outer
product of individual specific latent traits.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A real valued scalar. The contribution of the nominal trait to the
  ELBO.
}
\description{
Compute the contribution of a nominal trait to the Evidence Lower Bound
(ELBO) of a PLVM given the approximate posterior distribution for loadings
and latent traits.
}
