% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{compute_ordinal_auxiliary_trait_elbo}
\alias{compute_ordinal_auxiliary_trait_elbo}
\title{Ordinal Trait ELBO}
\usage{
compute_ordinal_auxiliary_trait_elbo(
  y,
  cut_off_points,
  loading_expectation,
  latent_trait_expectation,
  loading_outer_expectation,
  latent_trait_outer_expectation,
  perform_checks = TRUE
)
}
\arguments{
\item{y}{An N-dimensional vector of ordered factors with K levels. An ordinal
manifest trait.}

\item{cut_off_points}{A K+1 dimensional ordered vector of values. The cut off
points separating auxiliary traits into ordinal state.}

\item{loading_expectation}{A L-dimensional vector of real numbers, The row of
the expected loading matrix corresponding to the ordinal trait.}

\item{latent_trait_expectation}{An NxL matrix of real values. The expected
individual specific latent traits.}

\item{loading_outer_expectation}{A LxL matrix. The expected outer product for
the row of the expected loading matrix corresponding to the ordinal trait.}

\item{latent_trait_outer_expectation}{A LxLxN array, The expected outer
product of individual specific latent traits.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A real valued scalar. The contribution of the ordinal trait to the
  ELBO.
}
\description{
Compute the contribution of an ordinal trait to the Evidence Lower Bound
(ELBO) of a PLVM given the approximate posterior distribution for loadings
and latent traits.
}
