% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo.R
\name{compute_plvm_elbo}
\alias{compute_plvm_elbo}
\title{PLVM ELBO}
\usage{
compute_plvm_elbo(
  plvm_list,
  n_samples = 1000,
  random_seed = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{plvm_list}{A list defining the Phylogenetic Latent variable Model}

\item{n_samples}{A positive integer. The number of independent samples drawn
to obtain the Monte Carlo estimate.}

\item{random_seed}{A single value, interpreted as an integer, or NULL.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A scalar. The Evidence Lower Bound for a given Phylogenetic Latent
  Variable Model
}
\description{
Compute the Evidence Lower Bound (ELBO) for a specific a PLVM.Ê
}
\seealso{
initialise_plvm
}
