% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_loading.R
\name{compute_scaled_conditional_row_variance}
\alias{compute_scaled_conditional_row_variance}
\title{Compute the Scaled Conditional Row Variance of the Loading}
\usage{
compute_scaled_conditional_row_variance(
  loading_prior_correlation,
  perform_checks = TRUE
)
}
\arguments{
\item{loading_prior_correlation}{A D'xD' correlation matrix. The correlation
within loadings under the Gaussian prior.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A D'-dimensional vector of positive real numbers. The scaled
  conditional variance of each row in the loading matrix.
}
\description{
Computes the variance of the loading in each row conditioned on the loading
at every other row before allowing for the ARD precision hyper-parameter.
}
