% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_latent_traits.R
\name{compute_taxon_specific_latent_trait_elbo}
\alias{compute_taxon_specific_latent_trait_elbo}
\title{Taxon Specific Latent Trait Trait ELBO}
\usage{
compute_taxon_specific_latent_trait_elbo(
  taxon_specific_latent_trait_expectation,
  taxon_specific_latent_trait_outer_product_expectation,
  taxon_specific_latent_trait_covariance,
  phy,
  phylogenetic_gp,
  perform_checks = TRUE
)
}
\arguments{
\item{taxon_specific_latent_trait_expectation}{A (2S - 1)xL matrix of real
values. The expected value of the taxon specific latent traits.}

\item{taxon_specific_latent_trait_outer_product_expectation}{A LxLx(2S - 1)
array of real values. The expected outer product of taxon specific latent
traits under the approximate posterior.}

\item{taxon_specific_latent_trait_covariance}{A LxLx(2S - 1) array of real
values. The covariance of taxon specific latent traits under the
approximate posterior.}

\item{phy}{An object of class "\eqn{phylo}".}

\item{phylogenetic_gp}{A (2S - 1)x2xL array of real values. Hyperparameters
for the Gaussian process over \eqn{phy}.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A real valued scalar. The contribution of taxon specific latent
  traits to the ELBO.
}
\description{
Compute the contribution of taxon specific latent traits to the Evidence
Lower Bound (ELBO) of a PLVM given the approximate posterior distribution for
loadings and latent traits
}
