% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_latent_traits.R
\name{compute_terminal_taxon_specific_latent_trait_expectation}
\alias{compute_terminal_taxon_specific_latent_trait_expectation}
\title{Compute Terminal Taxon-specific Latent Trait Expectation}
\usage{
compute_terminal_taxon_specific_latent_trait_expectation(
  individual_specific_latent_traits,
  within_taxon_standard_deviation,
  parent_taxon_latent_trait,
  conditional_expectation_weight,
  conditional_standard_deviation,
  latent_trait_precision,
  perform_checks = TRUE
)
}
\arguments{
\item{individual_specific_latent_traits}{An L-dimensional vector of real
values. The latent trait values associated with individuals within the
taxon.}

\item{within_taxon_standard_deviation}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The within-taxon variation parameters for
each of the L latent traits.}

\item{parent_taxon_latent_trait}{An L-dimensional vector of real values. The
latent trait values associated with the parent taxon.}

\item{conditional_expectation_weight}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The weight of the latent trait values for the
parent taxon in the conditional distribution for the taxon-specific latent
trait.}

\item{conditional_standard_deviation}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The standard deviation of the taxon-specific
latent trait conditional on its parent taxon.}

\item{latent_trait_precision}{An L-dimensional vector of positive real
values. The precision of the taxon-specific latent trait.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An L-dimensional vector of real values.
}
\description{
Computes the expected value of taxon-specific latent traits at terminal nodes
of the phylogeny under the approximate posterior distribution for the PLVM.
}
