% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_latent_traits.R
\name{compute_terminal_taxon_specific_latent_trait_precision}
\alias{compute_terminal_taxon_specific_latent_trait_precision}
\title{Compute Terminal Taxon-specific Latent Trait Precision}
\usage{
compute_terminal_taxon_specific_latent_trait_precision(
  N_s,
  within_taxon_amplitude,
  conditional_standard_deviation,
  perform_checks = TRUE
)
}
\arguments{
\item{N_s}{A natural number. The number of individuals within the taxon.}

\item{within_taxon_amplitude}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The within-taxon variation parameters for
each of the L latent traits.}

\item{conditional_standard_deviation}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The standard deviation of the taxon-specific
latent trait conditional on its parent taxon.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An L-dimensional vector of real values. Precision parameters for the
  taxo-specific latent traits.
}
\description{
Computes the precision of taxon-specific latent traits at terminal nodes of
the phylogeny under the approximate posterior distribution for the PLVM.
}
