% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{default_skeleton}
\alias{default_skeleton}
\title{Default Skeleton VBScript}
\usage{
default_skeleton(print = FALSE)
}
\arguments{
\item{print}{If \code{TRUE}, print the file contents to the console.}
}
\value{
The path to the default skeleton VBScript.
}
\description{
Retrieve the path to the default skeleton VBScript file.
}
\details{
The skeleton VBScript must use the following placeholders:
\itemize{
\item \code{{macro_application}}: Placeholder for the macro application name,
e.g., "Word", "Excel", etc.
\item \code{{collection_type}}: Placeholder for the collection type
containing the macro, e.g., "Documents" for Word,
"Workbooks" for Excel, etc.
\item \code{{macro_file}}: Placeholder for the path to the file containing
the macro.
\item \code{{macro_name}}: Placeholder for the name of the macro.
\item \code{{macro_args}}: Placeholder for the macro arguments.
}
}
