% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{exp_quad_kernel}
\alias{exp_quad_kernel}
\title{The Exponentiated Quadratic kernel}
\usage{
exp_quad_kernel(d, amplitude, length_scale, perform_checks = TRUE)
}
\arguments{
\item{d}{A set of non-negative real numbers. Pairwise distances between
points.}

\item{amplitude}{A positive real number. The amplitude of the
Gaussian process.}

\item{length_scale}{A positive real number. The length-scale of the
Gaussian process.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An array of positive values between 0 and \eqn{amplitude} with the
  same dimension as d.
}
\description{
Computes \deqn{amplitude^2 * \exp ( - d^2 / 2 * length_scale )} where \eqn{d \ge 0}
is the distance between any two points in space for \eqn{amplitude > 0} and
\eqn{length_scale > 0}.
}
\examples{
x <- 1:10
d <- abs(outer(x, x, "-"))
K <- exp_quad_kernel(d, amplitude = 1, length_scale = 2)
}
