% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_generic_functions.R
\name{gaussian_outer_product_expectation}
\alias{gaussian_outer_product_expectation}
\title{Compute outer product expectation}
\usage{
gaussian_outer_product_expectation(
  expected_value,
  covariance_matrix,
  perform_checks = TRUE
)
}
\arguments{
\item{expected_value}{A real-valued vector length \eqn{L}. The expected value
of the Gaussian random variable.}

\item{covariance_matrix}{A \eqn{L \times L} covariance matrix. The covariance
matrix for the Gaussian random variable.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A symmetric \eqn{L \times L} matrix.
}
\description{
Computes the expected value of the outer product of the \eqn{D}-dimensional
Gaussian random variable.
}
\examples{
D <- 5
z <- rnorm(D)
S <- stats::rWishart(1, df = D, Sigma = diag(D)/ D^2)[, , 1]
gaussian_outer_product_expectation(expected_value = z, covariance_matrix = S)
}
