% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_updates.R
\name{heritable_amplitude_objective}
\alias{heritable_amplitude_objective}
\title{Objective function for heritable amplitude}
\usage{
heritable_amplitude_objective(
  par,
  i,
  heritable_amplitude,
  length_scale,
  taxon_specific_latent_trait_expectation,
  taxon_specific_latent_trait_outer_product_expectation,
  taxon_specific_latent_trait_covariance,
  phy,
  phylogenetic_gp,
  perform_checks = TRUE
)
}
\arguments{
\item{par}{A real value on the unit interval. The heritable amplitude.}

\item{i}{An integer value between 1 and L. The index for the heritable
amplitude to be optimised.}

\item{heritable_amplitude}{An L-dimensional vector of values on the unit
interval. The current amplitude of the phylogenetic OU process. Note that
this is the square root of the variance and so should be thought of as a
standard deviation parameter.}

\item{length_scale}{A positive real-valued scalar. The amplitude of the
phylogenetic OU process. Note that this is the square root of the variance
and so should be thought of as a standard deviation parameter.}

\item{taxon_specific_latent_trait_expectation}{A (2S - 1)xL matrix of real
values. The expected value of the taxon specific latent traits.}

\item{taxon_specific_latent_trait_outer_product_expectation}{A LxLx(2S - 1)
array of real values. The expected outer product of taxon specific latent
traits under the approximate posterior.}

\item{taxon_specific_latent_trait_covariance}{A LxLx(2S - 1) array of real
values. The covariance of taxon specific latent traits under the
approximate posterior.}

\item{phy}{An object of class "\eqn{phylo}".}

\item{phylogenetic_gp}{A (2S - 1)x2xL array of real values. Hyperparameters
for the Gaussian process over \eqn{phy}.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A scalar. The negative ELBO associated with individual specific
  latent traits scaled to account for the logit transform of the heritable
  amplitude.
}
\description{
Update the heritable amplitude by optimising the ELBO for taxon-specific
latent traits.
}
