% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{initialise_auxiliary_traits}
\alias{initialise_auxiliary_traits}
\title{Initialise Auxiliary Traits}
\usage{
initialise_auxiliary_traits(
  n_traits,
  manifest_trait_df,
  trait_names,
  trait_type,
  trait_levels,
  manifest_trait_index,
  auxiliary_trait_index,
  inverse_link_functions,
  auxiliary_traits = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{n_traits}{A natural number. The number of manifest traits.}

\item{manifest_trait_df}{The data frame of manifest traits. If included
allows for a non-exhaustive series of checks comparing metadata provided to
that implicit in the manifest traits.}

\item{trait_names}{A P-dimensional character vector. The manifest trait
names.}

\item{trait_type}{A P-dimensional vector of unordered factors belonging to
one of 4 levels. The trait type corresponding to each of the named manifest
traits. Ordinal traits are denoted \eqn{ord}, nominal are \eqn{nom},
scalar-valued continuous are \eqn{con}, and function-valued traits are
\eqn{fvt}.}

\item{trait_levels}{A P-dimensional vector. The number of levels associated with each
manifest trait. NA for continuous traits.}

\item{manifest_trait_index}{A P-dimensional list. columns of
\eqn{manifest_trait_df} associated with each trait. List elements will be
scalar valued for ordinal, nominal, and scalar-valued continuous traits.
Function-valued traits are multi-variate and indexed by a vector of values.}

\item{auxiliary_trait_index}{A P-dimensional list. Columns of the auxiliary
trait matrix to be associated with each trait. List elements will be scalar
valued for ordinal, nominal, and scalar-valued continuous traits.
Function-valued traits are multi-variate and indexed by a vector of values.}

\item{inverse_link_functions}{A P-dimensional list. Functions mapping
manifest traits to auxiliary traits. For traits taking continuous values
these must be specified by the user and can be the identity function.}

\item{auxiliary_traits}{An NxD' matrix of real numbers. The initial matrix of
auxiliary traits. When non null checks that the auxiliary trait matrix is
of the correct type and size.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An NxD' matrix of real numbers. The initial matrix of auxiliary
  traits.
}
\description{
Initialise auxiliary traits in a PLVM given manifest traits and metadata.
}
