% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_latent_traits.R
\name{initialise_individual_specific_latent_traits}
\alias{initialise_individual_specific_latent_traits}
\title{Initialise the Individual-specific latent traits}
\usage{
initialise_individual_specific_latent_traits(
  auxiliary_traits,
  loading,
  precision = 1,
  perform_checks = TRUE
)
}
\arguments{
\item{auxiliary_traits}{An NxD' matrix of real numbers. The initial matrix of
auxiliary traits.}

\item{loading}{A DxL dimensional matrix of real values. The PLVM loading.}

\item{precision}{A positive real valued scalar. A precision value allowing the
latent variables to be initialised. Larger imply that auxiliary traits are
more precisely observed.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An NxL dimensional matrix of real-valued latent traits.
}
\description{
Initialises individual specific latent traits given the auxiliary traits, a
set of loadings, and a vector of precision parameters associated with each
multivariate trait.
}
