% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_loading.R
\name{initialise_loading}
\alias{initialise_loading}
\title{Initialis Loading}
\usage{
initialise_loading(
  D_prime,
  L,
  ard_precision,
  loading_prior_correlation,
  loading = NULL,
  method = "random",
  random_seed = NULL,
  auxiliary_traits = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{D_prime}{An integer. The dimension of auxiliary traits traits.}

\item{L}{An integer. The dimension of latent traits.}

\item{ard_precision}{An L-dimensional vector of real numbers. The initial ARD
precision hyper-parameters on the loadings.}

\item{loading_prior_correlation}{A D'xD' correlation matrix. The correlation
within loadings under the Gaussian prior.}

\item{loading}{A D'xL matrix of real values. The initial loading matrix. When
non null checks that the loading matrix is of the correct type and size.}

\item{method}{A string. Either \eqn{"random"} or \eqn{"pca"} indicating the
type of initialisation adopted.}

\item{random_seed}{A single value, interpreted as an integer, or NULL.}

\item{auxiliary_traits}{An NxD' matrix of real numbers. The initial matrix of
auxiliary traits. Only required when \eqn{method == "pca"}}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A D'xL matrix of real values. The initial loading matrix.
}
\description{
Initialise loadings for a PLVM. Allows for either a random initialisation,
which does not depend on the auxiliary traits, or initialises with the first
\eqn{L} principle components, which does.
}
