% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_loading.R
\name{initialise_loading_ard_precision}
\alias{initialise_loading_ard_precision}
\title{Initialise ARD Precision}
\usage{
initialise_loading_ard_precision(
  L,
  ard_prior_shape = 1,
  ard_prior_rate = 1,
  ard_precision = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{L}{An integer. The dimension of latent traits.}

\item{ard_prior_shape}{A positive real-valued scalar. The shape of the Gamma
prior on the ARD precision hyper-parameters}

\item{ard_prior_rate}{A positive real-valued scalar. The rate of the Gamma
prior on the ARD precision hyper-parameters}

\item{ard_precision}{An L-dimensional vector of real numbers. The initial ARD
precision hyper-parameters on the loadings. When non-null checks that
\eqn{ard_precision} is of the correct type and size.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An L-dimensional vector of real numbers. The initial ARD precision
  hyper-parameters on the loadings
}
\description{
Initialiase the Automatic Relevance Determination hyper-parameter in the
Gaussian process prior on PLVM loadings.
}
