% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{initialise_precision}
\alias{initialise_precision}
\title{Initialise Precision}
\usage{
initialise_precision(
  n_traits,
  trait_names,
  trait_type,
  precision_prior_shape = 1,
  precision_prior_rate = 0.01,
  precision = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{n_traits}{A natural number. The number of manifest traits.}

\item{trait_names}{A P-dimensional character vector. The manifest trait
names.}

\item{trait_type}{A P-dimensional vector of unordered factors belonging to
one of 4 levels. The trait type corresponding to each of the named manifest
traits. Ordinal traits are denoted \eqn{ord}, nominal are \eqn{nom},
scalar-valued continuous are \eqn{con}, and function-valued traits are
\eqn{fvt}.}

\item{precision_prior_shape}{A positive real-valued scalar. The shape of the
Gamma prior on the precision.}

\item{precision_prior_rate}{A positive real-valued scalar. The rate of the
Gamma prior on the precision.}

\item{precision}{A vector length \eqn{n_traits} taking positive real values.
The Precision with which auxiliary traits are observed. Precision
parameters for discrete traits are fixed at 1. When non-null checks that
the precision parameters have been specified correctly.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A vector length \eqn{n_traits} taking positive real values. The
  Precision with which auxiliary traits are observed. Precision parameters
  for discrete traits are fixed at 1.
}
\description{
Initialise the precision parameters on auxiliary traits within the PLVM.
}
