% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_latent_traits.R
\name{latent_trait_outer_product_expectation}
\alias{latent_trait_outer_product_expectation}
\title{Compute latent trait outer product expectations}
\usage{
latent_trait_outer_product_expectation(z_n_tilde, S_n_z, perform_checks = TRUE)
}
\arguments{
\item{z_n_tilde}{A real-valued vector length \eqn{L}. The expected latent trait under the approximate posterior distribution.}

\item{S_n_z}{A \eqn{L \times L} covariance matrix. The variance of the latent trait under the approximate posterior distribution.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A symmetric \eqn{L \times L} matrix.
}
\description{
Computes the expected value of the outer product of the \eqn{L} latent traits for the \eqn{n^{th}} individual under the approximate posterior distribution.
}
\examples{
L <- 5
z <- rnorm(L)
S <- stats::rWishart(1, df = L, Sigma = diag(L)/ L^2)[, , 1]
latent_trait_outer_product_expectation(z_n_tilde = z, S_n_z = S)
}
