% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{macro_script}
\alias{macro_script}
\title{Build Macro Script}
\usage{
macro_script(macro_file, macro_name, ..., .skeleton = default_skeleton())
}
\arguments{
\item{macro_file}{The file containing the macro. Currently, ".xls*"
and ".doc*" files are supported.}

\item{macro_name}{The name of the macro.}

\item{...}{Arguments to pass to the macro. Type conversions in the
resulting VBScript are defined based on the R class of each
argument as follows:
\itemize{
\item numeric: cast to double using \code{Cdbl()}
\item integer: cast to integer using \code{CInt()}
\item character: cast to string using \code{CStr()}
}}

\item{.skeleton}{Path to the skeleton VBScript.
See \code{\link[=default_skeleton]{default_skeleton()}} for more information.}
}
\value{
A character string containing VBScript text.
}
\description{
Build a VBS script that calls the macro specified macro.
}
\examples{
example_file = system.file("examples", "data_importer.xlsm",
  package = "vbar")
macro_script(example_file, "importData",
  inputFile = NA_character_, targetSheet = NA_character_,
  targetRange = NA_character_, outputFile = NA_character_)

}
