% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{map_precision_to_auxiliary_traits}
\alias{map_precision_to_auxiliary_traits}
\title{Map Precision Parameters to Auxiliary Traits}
\usage{
map_precision_to_auxiliary_traits(
  precision,
  auxiliary_trait_index,
  perform_checks = TRUE
)
}
\arguments{
\item{precision}{A vector length \eqn{n_traits} taking positive real values.
The Precision with which auxiliary traits are observed. Precision
parameters for discrete traits are fixed at 1.}

\item{auxiliary_trait_index}{A P-dimensional list. Columns of the auxiliary
trait matrix to be associated with each trait. List elements will be scalar
valued for ordinal, nominal, and scalar-valued continuous traits.
Function-valued traits are multi-variate and indexed by a vector of values.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A D'-dimensional vector of positive real values. The precision along each dimension of the auxiliary traits.
}
\description{
Create a vector of precision parameters corresponding to the puxiliary traits.
}
