% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\name{nominal_inverse_link}
\alias{nominal_inverse_link}
\title{Nominal Inverse Link Function}
\usage{
nominal_inverse_link(
  y,
  mu,
  n_samples = 1000,
  return_expectation = TRUE,
  random_seed = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{y}{An N-dimensional vector of unordered factors with K levels. A nominal
manifest trait.}

\item{mu}{An N-dimensional vector of real numbers. The expected value of the
underlying Gaussian random variable with variance 1.}

\item{n_samples}{A positive integer. The number of independent samples drawn
to obtain the Monte Carlo estimate.}

\item{return_expectation}{Logical. Should manifest variables map to the
expected auxiliary variable or to values sampled from the truncated
Gaussian distribution.}

\item{random_seed}{A single value, interpreted as an integer, or NULL.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An NxK matrix of real values.
}
\description{
Map a nominal manifest trait to auxiliary traits.
}
