% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\name{nominal_link}
\alias{nominal_link}
\title{Nominal Link Function}
\usage{
nominal_link(X, levels, perform_checks = TRUE)
}
\arguments{
\item{X}{A NxK matrix of real values}

\item{levels}{An unordered factor vector of length K. The unordered
categories to which manifest traits can belong.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An N-dimensional vector of unordered factors with K levels. A nominal
  manifest trait.
}
\description{
Map the K-dimensional auxiliary trait to a nominal manifest trait.
}
