% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\name{nominal_probit_auxiliary_expectation}
\alias{nominal_probit_auxiliary_expectation}
\title{Expectation of multinomial probit auxiliary variables}
\usage{
nominal_probit_auxiliary_expectation(
  i,
  mu,
  n_samples = 1000,
  random_seed = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{i}{A positive integer. The index identifying the multinomial category..}

\item{mu}{A K-dimensional vector of real numbers. The expected value of the
auxiliary Gaussian mapping to the multinomial.}

\item{n_samples}{A positive integer. The number of independent samples drawn
to obtain the Monte Carlo estimate.}

\item{random_seed}{A single value, interpreted as an integer, or NULL.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A K-dimensional vector of real values.
}
\description{
Provides Monte Carlo estimates for the expected value of auxiliary variables in the
multinomial probit model given the manifest variable belongs in category \eqn{i} and the K-dimensional mean of the
underlying Gaussian distribution with variance 1.
}
