% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{nominal_probit_normalising_constant}
\alias{nominal_probit_normalising_constant}
\title{Nominal Probit Normalising Constant}
\usage{
nominal_probit_normalising_constant(
  i,
  mu,
  n_samples = 1000,
  random_seed = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{i}{A positive integer. The index identifying the multinomial category.}

\item{mu}{A K-dimensional vector of real numbers. The expected value of the
auxiliary Gaussian mapping to the multinomial.}

\item{n_samples}{A positive integer. The number of independent samples drawn
to obtain the Monte Carlo estimate.}

\item{random_seed}{A single value, interpreted as an integer, or NULL.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A real number on the unit interval.
}
\description{
Computes a Monte Carlo estimate for probability that a K-dimensional
multinomial belongs to category \eqn{i} under the probit model given the mean
of the underlying Gaussian distribution with variance 1.
}
