% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{ordinal_inverse_link}
\alias{ordinal_inverse_link}
\title{Ordinal Trait Inverse Link Function}
\usage{
ordinal_inverse_link(
  y,
  cut_off_points,
  mu,
  return_expectation = TRUE,
  random_seed = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{y}{An N-dimensional vector of ordered factors with K levels. An ordinal
manifest trait.}

\item{cut_off_points}{A K+1 dimensional ordered vector of values. The cut off
points separating auxiliary traits into ordinal state.}

\item{mu}{An N-dimensional vector of real numbers. The expected value of the
underlying Gaussian random variable with variance 1.}

\item{return_expectation}{Logical. Should manifest variables map to the
expected auxiliary variable or to values sampled from the truncated
Gaussian distribution.}

\item{random_seed}{A single value, interpreted as an integer, or NULL.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An N-dimensional vector of real numbers. An auxiliary trait.
}
\description{
Map an ordinal manifest trait to an auxiliary trait.
}
