% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\name{ordinal_link}
\alias{ordinal_link}
\title{Ordinal Trait Link function}
\usage{
ordinal_link(x, cut_off_points, perform_checks = TRUE)
}
\arguments{
\item{x}{An N-dimensional vector of real numbers. An auxiliary trait.}

\item{cut_off_points}{A K+1 dimensional ordered vector of values. The cut off
points separating auxiliary traits into ordinal state.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An N-dimensional vector of ordered factors with K levels. An ordinal
  manifest trait.
}
\description{
Map an auxiliary trait to ordinal manifest trait.
}
