% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{ordinal_probit_normalising_constant}
\alias{ordinal_probit_normalising_constant}
\title{Ordinal Probit Normalising Constant}
\usage{
ordinal_probit_normalising_constant(
  y,
  mu,
  cut_off_points,
  log_out = TRUE,
  perform_checks = TRUE
)
}
\arguments{
\item{y}{An N-dimensional vector of ordered factors with K levels. An ordinal
manifest trait.}

\item{mu}{An N-dimensional vector of real numbers. The expected value of the
underlying Gaussian random variable with variance 1.}

\item{cut_off_points}{A K+1 dimensional ordered vector of values. The cut off
points separating auxiliary traits into ordinal state.}

\item{log_out}{Logical. Should the log probability be returned.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A vector of (log) probabilities.
}
\description{
Computes the probability that ordinal variables belonging to one of K ordered
categories each belong to category \eqn{i} under the probit model given the
mean of the underlying Gaussian distribution with variance 1.
}
