% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic_hyperparameters.R
\name{phylogenetic_ou_hyperparameters}
\alias{phylogenetic_ou_hyperparameters}
\title{Phylogenetic Ornstein-Uhlenbeck hyperparameters}
\usage{
phylogenetic_ou_hyperparameters(
  phy,
  amplitude = 1,
  length_scale = 2,
  perform_checks = TRUE
)
}
\arguments{
\item{phy}{An object of class "\eqn{phylo}".}

\item{amplitude}{A positive real number. The amplitude of the
Ornstein-Uhlenbeck process.}

\item{length_scale}{A positive real number. The length-scale of the
Ornstein-Uhlenbeck process.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A \eqn{2n - 1 \times 2} matrix of positive real values, where \eqn{n}
  the the number of terminal nodes on a bifurcating tree. The first column is
  the contribution of the parent node to the expected value, while the second
  is the conditional variance given the parent.
}
\description{
Compute the hyper-parameters at each node on a phylogeny given the amplitude
and length-scale of the Ornstein-Uhlenbeck process.
}
\examples{
tr <- ape::rtree(100)
hyper <- phylogenetic_ou_hyperparameters(tr)
}
