% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparameterise_phylogenetic_ou.R
\name{reparameterise_phylogenetic_ou}
\alias{reparameterise_phylogenetic_ou}
\title{Reparameterise phylogenetic Ornstein-Uhlenbeck}
\usage{
reparameterise_phylogenetic_ou(
  phy,
  heritable_amplitude = 1,
  length_scale = 2,
  environmental_amplitude = 0,
  perform_checks = TRUE
)
}
\arguments{
\item{phy}{An object of class "\eqn{phylo}".}

\item{heritable_amplitude}{A positive real-valued scalar. The amplitude of
the phylogenetic OU process.}

\item{length_scale}{A positive real number. The length-scale of the
Ornstein-Uhlenbeck process.}

\item{environmental_amplitude}{A positive real-valued scalar. The amplitude
of non-phylogenetic variation in the traits of extant taxa.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A \eqn{2n - 1 \times 2} matrix of positive real values, where \eqn{n}
  the the number of terminal nodes on a bifurcating tree. The first column is
  the weight of the parent node to the expected value, while the second is
  the conditional standard deviation given the parent.
}
\description{
Reparameterise a phylogenetic OU process as a Gauss-Markov process over a
phylogeny.
}
\examples{
S <- 100
tr <- ape::rtree(S)
par <- reparameterise_phylogenetic_ou(tr)
}
