% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_phylo.R
\name{scale_phylo}
\alias{scale_phylo}
\title{Rescale a Phylogenetic tree}
\usage{
scale_phylo(phy, max_dist = 1)
}
\arguments{
\item{phy}{An object of class "\eqn{phylo}".}

\item{max_dist}{A positive, real scalar. The maximum distance along the
rescaled phylogeny from root to tip.}
}
\value{
An object of class "\eqn{phylo}" where the maximum distance from the
  root to tip is \eqn{max_dist}
}
\description{
Rescale branch lengths for a \eqn{phylo} object to fix the maximum distance
from the root to tip of a phylogenetic tree.
}
\examples{
tr <- ape::rtree(100)
tr1 <- scale_phylo(tr)
}
