% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{script_dir}
\alias{script_dir}
\title{Retrieve Session Script Directory}
\usage{
script_dir()
}
\value{
The path to the script directory.
}
\description{
Retrieve the script directory created for this session.
}
