% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phylogenetic_gaussian_process.R
\name{simulate_phylogenetic_ou}
\alias{simulate_phylogenetic_ou}
\title{Simulate a Phylogenetic Ornstein-Uhlenbeck process}
\usage{
simulate_phylogenetic_ou(
  phy,
  amplitude = 1,
  length_scale = 2,
  internal = FALSE,
  perform_checks = TRUE
)
}
\arguments{
\item{phy}{An object of class "\eqn{phylo}".}

\item{amplitude}{A positive real number. The amplitude of the
Ornstein-Uhlenbeck process.}

\item{length_scale}{A positive real number. The length-scale of the
Ornstein-Uhlenbeck process.}

\item{internal}{A logical value. Specifies whether to return the values at
internal nodes}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A numeric vector with names taken from the tip labels of \eqn{phy}. If
  \eqn{ancestor = TRUE}, the node labels are used if present, otherwise, “Node1”,
  “Node2”, etc.
}
\description{
A wrapper for \eqn{ape::rTraitCont} which simulates an Ornstein-Uhlenbeck
process parameterised by \eqn{amplitude > 0} and \eqn{length_scale > 0}.
}
\examples{
tr <- ape::rtree(100)
z <- simulate_phylogenetic_ou(tr)
}
