% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_trait_model_specification}
\alias{synthetic_trait_model_specification}
\title{Model Specification for Synthetic Traits}
\format{
A list of length 9:
\describe{
  \item{phylogeny}{An object of type \eqn{phylo} with \eqn{S = 128} terminal
   taxa and \eqn{S-1} internal taxa}
   \item{auxiliary_traits}{A named 384x37 matrix of real numbers. Auxiliary
   traits mapping to synthetic manifest traits.}
   \item{loading}{A named 37x4 matrix of real numbers}
   \item{auxiliary_trait_precision}{A 4-dimensional vector of positive real
   numbers. Precision with which auxiliary traits are observed.}
   \item{individual_specific_latent_traits}{A 384x3 matrix of real numbers.
   The latent trait associated with each observed manifest trait.}
   \item{taxon_specific_latent_traits}{A 255x3 matrix of real numbers.
   The latent trait associated with each taxa.}
   \item{within_taxon_sd}{A 3 dimensional vector of real values on the
   interval (0, 1). Standard deviation of the within-taxon variation.}
   \item{heritability}{A 3 dimensional vector of real values on the
   interval (0, 1). The heritability of latent traits over the phylogeny.}
   \item{ordinal_trait_cut_off}{A 5-dimensional vector of increasing values.
   Cut-off points for the ordinal trait mapping}
}
}
\usage{
synthetic_trait_model_specification
}
\description{
The Phylogenetic Latent Variable Model (PLVM) specified in the package
README. Used to illustrate the approach to ancestral reconstruction offered
by Variational Inference for the PLVM.
}
\keyword{datasets}
