% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_traits}
\alias{synthetic_traits}
\title{Synthetic Phenotypic Traits}
\format{
A data frame with 384 observations of 36 variables:
\describe{
  \item{taxon_id}{The taxon to which the individual belongs.}
  \item{ord}{An
  ordinal trait for which K = 4}
  \item{nom}{A nominal trait for which K = 3}
  \item{con}{A scalar-valued continuous trait.}
  \item{fvt}{A function-valued trait lying on the positive real
  numbers, observed at 32 points registered along the continuum.}
}
}
\usage{
synthetic_traits
}
\description{
A set of synthetic manifest traits generated sampled from a Phylogenetic
Latent Variable Model (PLVM). Used to illustrate the approach to ancestral
reconstruction offered by Variational Inference for the PLVM.
}
\keyword{datasets}
