% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_auxiliary_traits.R
\name{update_discrete_auxiliary_traits}
\alias{update_discrete_auxiliary_traits}
\title{Update Discrete Auxiliary Traits}
\usage{
update_discrete_auxiliary_traits(
  manifest_trait_df,
  metadata,
  auxiliary_traits,
  loading_expectation,
  latent_trait_expectation,
  n_samples = 1000,
  random_seed = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{manifest_trait_df}{The data frame of manifest traits. If included
allows for a non-exhaustive series of checks comparing metadata provided to
that implicit in the manifest traits.}

\item{metadata}{A data frame. Contains all the metadata required to map a set
of manifest traits to the PLVM auxiliary traits..}

\item{auxiliary_traits}{An NxD' matrix of real numbers. The auxiliary traits.}

\item{loading_expectation}{A D'xL matrix of real numbers, The expected
loading matrix.}

\item{latent_trait_expectation}{An NxL matrix of real values. The expected
individual specific latent traits.}

\item{n_samples}{A positive integer. The number of independent samples drawn
to obtain the Monte Carlo estimate.}

\item{random_seed}{A single value, interpreted as an integer, or
NULL.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
An NxD' Matrix of real values. The updated auxiliary traits.
}
\description{
Update the auxiliary traits corresponding to discrete manifest traits given
the CAVI aprroximate posterior over the loading and latent traits in the
PLVM.
}
