% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_updates.R
\name{update_loading_ard_precision}
\alias{update_loading_ard_precision}
\title{Update ARD loading precision}
\usage{
update_loading_ard_precision(
  loading_col_outer_product_expectation,
  inv_loading_prior_correlation,
  ard_shape = 1,
  ard_rate = 1
)
}
\arguments{
\item{loading_col_outer_product_expectation}{A D'xD'xL array of real values.
The expected outer product for each column of the loading matrix under the
approximate posterior distribution.}

\item{inv_loading_prior_correlation}{A D'xD' matrix. The inverse of the prior
correlation matrix which defines the relationship between elements of the
loading matrix.}

\item{ard_shape}{A positive real-valued scalar. The shape of the Gamma
distribution generating initial ARD precision hyper-parameters}

\item{ard_rate}{A positive real-valued scalar. The rate of the Gamma
distribution generating initial ARD precision hyper-parameters}
}
\value{
An L-dimensional vector. The ARD precision associated with each
  column of the loading matrix.
}
\description{
Update the ARD precision associated with each loading within the PLVM by
optimising the ELBO.
}
