% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_updates.R
\name{update_precision}
\alias{update_precision}
\title{Update auxiliary trait precision}
\usage{
update_precision(
  precision,
  metadata,
  auxiliary_traits,
  loading_expectation,
  latent_trait_expectation,
  loading_outer_expectation,
  latent_trait_outer_expectation
)
}
\arguments{
\item{precision}{A P-dimensional vector of positive real values. The
precision with which auxiliary traits are observed.}

\item{metadata}{A data frame. Contains all the metadata required to map a set
of manifest traits to the PLVM auxiliary traits..}

\item{auxiliary_traits}{An NxD' matrix of real numbers. The auxiliary traits.}

\item{loading_expectation}{A D'xL matrix of real numbers, The expected
loading matrix.}

\item{latent_trait_expectation}{An NxL matrix of real values. The expected
individual specific latent traits.}

\item{loading_outer_expectation}{A LxLxD' array. The expected outer products
of the expected loading matrix.}

\item{latent_trait_outer_expectation}{A LxLxN array. The expected outer
product of individual specific latent traits.}
}
\value{
A P-dimensional vector. The precision associated with each auxiliary
  trait.
}
\description{
Update the precision associated with each auxiliary trait within the PLVM by
optimising the ELBO.
}
