% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_updates.R
\name{within_taxon_amplitude_objective}
\alias{within_taxon_amplitude_objective}
\title{Objective function for within-taxon amplitude}
\usage{
within_taxon_amplitude_objective(
  par,
  i,
  individual_specific_latent_trait_expectation,
  taxon_id,
  phy,
  terminal_taxon_specific_latent_trait_expectation,
  individual_specific_latent_trait_covariance,
  individual_specific_latent_trait_outer_product_expectation,
  terminal_taxon_latent_trait_outer_product_expectation,
  within_taxon_amplitude,
  perform_checks = TRUE
)
}
\arguments{
\item{par}{A real value on the unit interval. The within-taxon amplitude.}

\item{i}{An integer value between 1 and L. The index for the within-taxon
amplitude to be optimised.}

\item{individual_specific_latent_trait_expectation}{A NxL matrix of real
values. The expected value of the individual specific latent traits.}

\item{taxon_id}{A N-dimensional vector of labels. Matches each individual
specific latent trait to a terminal node in \eqn{phy}. Each element must
correspond to one of \eqn{phy$tip.labels}.}

\item{phy}{An object of class "\eqn{phylo}".}

\item{terminal_taxon_specific_latent_trait_expectation}{A SxL matrix of real
values. The expected taxon-specific latent trait for terminal nodes in
\eqn{phy}. Rows must be ordered by \eqn{phy$tip.labels}.}

\item{individual_specific_latent_trait_covariance}{A LxL covariance matrix.
The covariance of individual specific latent traits under the approximate
posterior.}

\item{individual_specific_latent_trait_outer_product_expectation}{A LxLxN
array of real values. The expected outer product of individual specific
latent traits under the approximate posterior.}

\item{terminal_taxon_latent_trait_outer_product_expectation}{A LxLxS array of
real values. The expected outer product of taxon specific latent traits at
terminal nodes under the approximate posterior.}

\item{within_taxon_amplitude}{An L-dimensional vector of real values
on the interval \eqn{[0, 1]}. The within-taxon variation parameters for
each of the L latent traits.}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A scalar. The negative ELBO associated with individual specific
  latent traits scaled to account for the logit transform of the within-taxon
  amplitude.
}
\description{
Update the within-taxon amplitude by optimising the ELBO for
individual-specific latent traits.
}
