% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phylogenetic_gaussian_process.R
\name{simulate_phylogenetic_ou}
\alias{simulate_phylogenetic_ou}
\title{Simulate a Phylogenetic Ornstein-Uhlenbeck process}
\usage{
simulate_phylogenetic_ou(
  phy,
  heritable_amplitude = 1,
  length_scale = 2,
  environmental_amplitude = 0,
  internal = FALSE,
  perform_checks = TRUE
)
}
\arguments{
\item{phy}{An object of class "\eqn{phylo}".}

\item{heritable_amplitude}{A positive real-valued scalar. The amplitude of
the phylogenetic OU process.}

\item{length_scale}{A positive real number. The length-scale of the
Ornstein-Uhlenbeck process.}

\item{environmental_amplitude}{A positive real-valued scalar. The amplitude
of non-phylogenetic variation in the traits of extant taxa.}

\item{internal}{A logical value. Specifies whether to return the values at
internal nodes}

\item{perform_checks}{Logical. Check if function inputs are specified
correctly.}
}
\value{
A numeric vector with names taken from the tip labels of \eqn{phy}.
  If \eqn{ancestor = TRUE}, the node labels are used if present, otherwise,
  “Node1”, “Node2”, etc.
}
\description{
A wrapper for \eqn{ape::rTraitCont} which simulates an Ornstein-Uhlenbeck
process over the phylogeny \eqn{phy} such that the phylogenetic Gaussian
process \deqn{ f ( \boldsymbol t )) \sim \mathcal{GP} ( 0, k (t, t' ))))} for
\deqn{ k (t, t' )) = heritable_amplitude^2 \exp ( - \frac{d (t, t'
))}{length_scale})) + environmental_amplitude^2 \delta ( t_1, \dots, t_S )) }
parameterised by \eqn{heritable_amplitude > 0}, \eqn{length_scale > 0}, and
\eqn{environmental_amplitude > 0}, where \eqn{t} is a position on the
phylogeny, \eqn{d (t, t')} is the length of the shortest path from \eqn{t} to
\eqn{t'}, \eqn{t_1, \dots, t_S} are positions corresponding to the \eqn{S}
extant taxa, and \eqn{\delta ( \cdot ))} is the indicator function.
}
\examples{
S <- 100
tr <- ape::rtree(S)
z <- simulate_phylogenetic_ou(tr)
}
